requirements.txt
anyio==4.3.0
black==23.9.1
blinker==1.7.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dnspython==2.6.1
filelock==3.12.4
Flask==3.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.30
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.11.2
phonenumbers==8.13.33
platformdirs==3.11.0
pre-commit==3.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.9
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
requests==2.31.0
ruff==0.0.292
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
truecallerpy==1.0.3
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.24.5
wcwidth==0.2.13
Werkzeug==3.0.1
